Refactor task into submodules
